% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{alphaifCalc}
\alias{alphaifCalc}
\title{alphaifCalc}
\usage{
alphaifCalc(pxV, pxCenterV, samDet, tiltV = 0, pxSizeV = 0.172)
}
\arguments{
\item{pxV}{pixel position}

\item{pxCenterV}{cordinate of direct beam position}

\item{samDet}{sample to detector distance}

\item{tiltV}{vertical detector tilt}

\item{pxSizeV}{detector pixel size in vertical direction}
}
\value{
angle alpha.i + alpha.f for pxV in deg.
}
\description{
calculate angle alpha.i + alpha.f from from known gisaxs
             parameter calibration (central pixel, sample to detector distance)
}
\examples{
alphaifCalc(100, 50, 1700, -5.21)
}
\author{
K. Juraic
}
