% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisaxs_model.R
\name{alphaT}
\alias{alphaT}
\title{alphaT(alpha, delta, beta)}
\usage{
alphaT(ai, delta, beta)
}
\arguments{
\item{ai}{angle of incidence in [deg.]}

\item{delta}{Re part of index of refraction}

\item{beta}{Im part of index of refraction}
}
\value{
at angle of refracted beam in [deg.]
}
\description{
calculate angle of refracted beam
}
\examples{
alphaT(ai = 0.5, delta = 7e-5, beta = 5e-7)
}
\author{
K. Juraic
}
