% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calibration.R
\name{alphaTcalc}
\alias{alphaTcalc}
\title{Angle of refracted beam calculation}
\usage{
alphaTcalc(ai, delta, beta)
}
\arguments{
\item{ai}{angle of incidence}

\item{delta}{Re index of refraction}

\item{beta}{Im. index of refracion}
}
\value{
alphaT angle of refracted beam
}
\description{
angle of transmited beam trough surface
}
\examples{
alphaTcalc(.22,0.758e-5, 1.755e-7)
}
\author{
K. Juraic
}

