% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{angle2Q}
\alias{angle2Q}
\title{angle2Q}
\usage{
angle2Q(ai, af, tthf, lambda = 0.154)
}
\arguments{
\item{ai}{angle of incidence in spec plane}

\item{af}{exit angle in spec plane}

\item{tthf}{2*theta exit angle}

\item{lambda}{wavelength}
}
\value{
Q vector(qx,qy,qz)
}
\description{
calculate Q vector from angle pair (alpha.f, tth.f)
}
\examples{
angle2Q(.22, .28, 0)
}
\author{
K. Juraic
}

