% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{constructLogBook}
\alias{constructLogBook}
\title{Read and insert data in GISAXS/GIWAXS logBook file}
\usage{
constructLogBook(dataDir, logFile = file.choose())
}
\arguments{
\item{dataDir}{folder with tif files}

\item{logFile}{name of logBook file}
}
\value{
logBook data.frame(name, time, angle, sample)
}
\description{
Read existing logBook files and insert additional data if
             neceseary. If file does not exist read all tif files from
             dataDir and manualy insert aditionay data. Refreshed logBook
             data.frame will be saved to file logFile. Function can
             automaticaly read exposure times from tif files. Data.frame
             logDf is dynamicaly added to global environment
}
\examples{
\dontrun{constructLogBook(dataDir = data_path_raw, logFile = logFile)}
}
\author{
Krunoslav Juraic
}
