% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_distributions.R
\name{dSZ}
\alias{dSZ}
\title{Shultz-Zimm distribution}
\usage{
dSZ(r, Rmean, Rsd)
}
\arguments{
\item{r}{x value for which calculation is done}

\item{Rmean}{distribution paramaterer (average)}

\item{Rsd}{distribution parameter (standard deviation)}
}
\value{
distribution density for given x
}
\description{
Calculate Shultz Size distribution density
             For particle size distribution is better to use Shultz-Zihm distribution
             then normal size distribution because its possible to avoid negative density values
             SOURCE: M. khaneft, phd thesis: polymers in aligned carbon nanotube arrays (2013)
}
\examples{
       dSZ(2.0, 1.0, 0.3)
}
\author{
K. Juraic
}

