% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/export_data.R
\name{export_pdh}
\alias{export_pdh}
\title{export SAXS data in pdh format}
\usage{
export_pdh(param_list, saxs_dat, file_name)
}
\arguments{
\item{param_list}{experiment parameters}

\item{saxs_dat}{experimental data}

\item{file_name}{output pdh file name}
}
\value{
none
}
\description{
export SAXS data in pdh format
}
\examples{
saxs_dat <- data.frame(q = 1:100, I = sqrt(1:100), Ierr = sqrt(1:100)*.1)
file_name <- "proba.phd"
param_list <- list(title = "AustroSAXS",
                   keywords = "SAXS",
                   number_of_points = nrow(saxs_dat),
                   sample_to_detector_mm = 1741.829,
                   normalization_factor = 1,
                   lambda_nm = 1.54e-1
)
   export_pdh(param_list, saxs_dat, file_name)
}

