% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_factor.R
\name{ff_cylinder_isgisaxs}
\alias{ff_cylinder_isgisaxs}
\title{Cylinder Form Factor (IsGISAXS)}
\usage{
ff_cylinder_isgisaxs(qx, qy, qz, R, H)
}
\arguments{
\item{qx}{x component of wave vector Q}

\item{qy}{y component of wave vector Q}

\item{qz}{z component of wave vector Q}

\item{R}{cylinder radius in nm}

\item{H}{cylinder height in nm}
}
\value{
Form Factor value
}
\description{
calculate form factor for cylinder (radius R and height H) for
             wave vector Q(qx, qy, qz). Cylinder Z axis is normal to sample 
             surface.
}
\examples{
   \dontrun{ff_cylinder_isgisaxs(.1, 1., .5, 10, 100)}
}
\references{
IsGISAXS manual
}
\author{
K. Juraic
}
