% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saxs_fit.R
\name{fGauss}
\alias{fGauss}
\title{fGauss(x, p0, pf, pc, pw)}
\usage{
fGauss(x, p0, pf, pc, pw)
}
\arguments{
\item{x}{x coord for which Gauss is calculated}

\item{p0}{shift factor}

\item{pf}{scaling factor}

\item{pc}{position of gaussian peak}

\item{pw}{width of gaussian peak}
}
\value{
y = Gaussina value at x
}
\description{
calculate Gaussian
}
\examples{
fGauss(.2,1,10,5,2)
}
\author{
K. Juraic
}
