% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/saxs_fit.R
\name{fitGauss}
\alias{fitGauss}
\title{fitGauss(x, y, subset)}
\usage{
fitGauss(x, y, subset = rep(TRUE, times = length(x)))
}
\arguments{
\item{x}{array of independent variable}

\item{y}{variable of dependent variable}

\item{subset}{TRUE/FALSE array which x,y pairs to include in fit}
}
\value{
fit fited parameters
}
\description{
fit Gaussian to x,y data with automatic determination of init parameters
             fit by using nlsLM from minpac.lm package
}
\examples{
fGauss(.2,1,10,5,2)
}
\author{
K. Juraic
}

