% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tif_utility.R
\name{gisaxs2D_angle}
\alias{gisaxs2D_angle}
\title{gisaxs2D_angle(mat, aif, tthf, tif_name, save2file = FALSE, width = 600, heiht = 400, title = TRUE)}
\usage{
gisaxs2D_angle(mat, aif, tthf, tif_name, save2file = FALSE, width = 2,
  height = 1, title = TRUE)
}
\arguments{
\item{mat}{z values (intensity)}

\item{aif}{alpha_if scale values}

\item{tthf}{tthf scale values}

\item{tif_name}{tif file name}

\item{save2file}{save to file}

\item{width}{image width}

\item{height}{image height}

\item{title}{to put title or not}
}
\description{
display 2D matrix data with Q scale and/or save in png file
}
\examples{
\dontrun{gisaxs2D_Q(mat, aif, tthf, tif_name,
                     save2file = FALSE,
                     width = 600, heiht = 400, title = TRUE)}
}
\author{
K. Juraic
}
