% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tif_utility.R
\name{gisaxs2D_Q}
\alias{gisaxs2D_Q}
\title{gisaxs2D_Q(mat, Qy, Qz, tif_name, save2file = FALSE, width = 600, heiht = 400, title = TRUE)}
\usage{
gisaxs2D_Q(mat, Qy, Qz, tif_name, save2file = FALSE, width = 600,
  height = 400, title = TRUE)
}
\arguments{
\item{mat}{z values (intensity)}

\item{Qy}{Qy scale values}

\item{Qz}{Qz scale values}

\item{tif_name}{tif file name}

\item{save2file}{save to file}

\item{width}{image width}

\item{height}{image height}

\item{title}{to put title or not}
}
\description{
display 2D matrix data with Q scale and/or save in png file
}
\examples{
\dontrun{gisaxs2D_Q(mat, Qy, Qz, tif_name, save2file = FALSE,
                     width = 600, heiht = 400, title = TRUE)}
}
\author{
K. Juraic
}

