% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/section.R
\name{lstSec2file}
\alias{lstSec2file}
\title{save list of section data.frame to file}
\usage{
lstSec2file(secLst, colSubset = 1:nrow(secLst[[1]]))
}
\arguments{
\item{secLst}{list of section data frame}

\item{colSubset}{subset of column to save to file}
}
\description{
save list of section data.frame to file
}
\examples{
\dontrun{lstSec2file(sec, fname)}
}

