% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\name{lstSecDfAddAngle}
\alias{lstSecDfAddAngle}
\title{Add angular scale to list of section}
\usage{
lstSecDfAddAngle(secLst, aiLst, calib)
}
\arguments{
\item{secLst}{list of section data frames (pxH, pxV, Imean, Isd)}

\item{aiLst}{list of angle of incidence in deg.}

\item{calib}{calibration parameters (pxcH, pxcV, saDet, tiltH, tiltV)}
}
\value{
seclst list of section data.frame(pxH, pxV, Imean, Isd, aif, ai, tthf)
}
\description{
Add angular scale columns (aif, af, tthf) to list of section data frame
}
\examples{
 \dontrun{lstSecDfAddAngle(secLst, aiLst, calib)}
}
\author{
K. Juraic
}

