% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\name{lstSecDfAddQ}
\alias{lstSecDfAddQ}
\title{Add Q scale to list of section}
\usage{
lstSecDfAddQ(secLst, aiLst, lambda = 0.154)
}
\arguments{
\item{secLst}{list of section data frame (pxH, pxV, Imean, Isd, aif, ai, tthf)}

\item{aiLst}{list of angle of incidence in deg.}

\item{lambda}{wavelength in nm}
}
\value{
secLst list of data.frame(pxH, pxV, Imean, Isd, aif, ai, tthf, Qx, Qy, Qz)
}
\description{
Add Q scale columns (qx, Qy, Qz) to list of section data frame
}
\examples{
 \dontrun{lstSecDfAddAngle(secLst, aiLst, lambda)}
}
\author{
K. Juraic
}
