% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tif_utility.R
\name{mat_max_pos}
\alias{mat_max_pos}
\title{position of maximum in part of matrix}
\usage{
mat_max_pos(mat, subset_H = 1:dim(mat)[1], subset_V = 1:dim(mat)[2],
  show = TRUE)
}
\arguments{
\item{mat}{2D matrix}

\item{subset_H}{horizontal subsection}

\item{subset_V}{vertical subsection}

\item{show}{display m}
}
\value{
c(h_max, v_max) cordinates of matrix
}
\description{
position of maximum in part of matrix
}
\examples{
\dontrun{mat_max_pos(mat)}
}

