% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tif_utility.R
\name{matCrop}
\alias{matCrop}
\title{matrix crop}
\usage{
matCrop(mat, bndRow, bndCol)
}
\arguments{
\item{mat}{matrix}

\item{bndRow}{c(px1,px2) horizontal boundaries for croping in pixels}

\item{bndCol}{c(px1,px2) vertical boundaries for croping in pixels}
}
\value{
list(pxRow, pxCol, mat) cropped matrix with croped horizontal and vertical scale
}
\description{
crop matrix to specific region
}
\examples{
m <- matrix(1:100,nrow = 10, ncol = 10, byrow = TRUE)
crop_px_h <- c(2,8)
crop_px_v <- c(3,6)
matCrop(m,crop_px_h,crop_px_v)
}
\author{
K. Juraic
}

