% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tif_utility.R
\name{matLstShow}
\alias{matLstShow}
\title{matLstShow(matLst, titles, ...)}
\usage{
matLstShow(matLst, titles, ...)
}
\arguments{
\item{matLst}{list of matrix data}

\item{titles}{main title for plots}

\item{...}{other parameters log='z'}
}
\description{
Show list of image matrix data
}
\examples{
\dontrun{matLstShow(dat, fNames, log='z')}
}
\author{
K. Juraic
}

