% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tif_utility.R
\name{matPlot}
\alias{matPlot}
\title{matPlot(z, x = 1:dim(z)[1], y = 1:dim(z)[2], scaleType = '', sampling = 1, ...)}
\usage{
matPlot(z, x = 1:dim(z)[1], y = 1:dim(z)[2], scaleType = "",
  sampling = 1, ...)
}
\arguments{
\item{z}{z values (intensity)}

\item{x}{x scale values}

\item{y}{y scale values}

\item{scaleType}{'q'= Q scale, 'a'=angular scale}

\item{sampling}{skip sampling pixels for fast drawing}

\item{...}{other parameters log='z', main='title'}
}
\description{
display 2D matrix data with specific scale
}
\examples{
m <- matrix(1:100,nrow = 10, ncol = 10, byrow = TRUE)
matPlot(m)
}
\author{
K. Juraic
}
