% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tif_utility.R
\name{matrix_shift}
\alias{matrix_shift}
\title{matrix shift}
\usage{
matrix_shift(m, shift)
}
\arguments{
\item{m}{matrix}

\item{shift}{c(h,v) number of columns and rows for shift}
}
\value{
m_shift shifted matrix
}
\description{
shift matrix rows and columns
}
\examples{
m <- matrix(1:100,nrow = 10, ncol = 10, byrow = TRUE)
shift <- c(0,-2)
matrix_shift(m,shift)
}
\author{
K. Juraic
}
