% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{penDepth}
\alias{penDepth}
\title{penDepth}
\usage{
penDepth(ai, delta, beta, lambda = 0.154)
}
\arguments{
\item{ai}{angle of incidence}

\item{delta}{Re part of index of refraction}

\item{beta}{Im. part of index of refraction}

\item{lambda}{wavelength [nm]}
}
\value{
calculated penetration depth in [nm]
}
\description{
penetration depth calculation
}
\examples{
ai <- seq(.01,1,by=.001)
plot(ai,penDepth(ai, 0.758e-5, 1.755e-7), type='l')
}
\author{
K. Juraic
}

