% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/porosity.R
\name{porosity}
\alias{porosity}
\title{Porosity estimation from crtical angle for total external reflectio}
\usage{
porosity(ac_exp, ac_bulk)
}
\arguments{
\item{ac_exp}{experimental critical angle for total external reflection}

\item{ac_bulk}{theoretical bulk critical angle for ideal material}
}
\value{
value for porosity [0-1]
}
\description{
Porosity estimation from crtical angle for total external reflectio
}
\examples{
 # Porosity calculation for amorphous silicon
 porosity(.18, .223)
}
\author{
K. Juraic
}

