% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import_data.R
\name{read_nika_sec}
\alias{read_nika_sec}
\title{Read Nika section file}
\usage{
read_nika_sec(fname)
}
\arguments{
\item{fname}{Nika section file name}
}
\value{
data.frame(Q, Qy, Qz, Intensity, Error)
}
\description{
read horizontal/vertical section od 2D SAXS images done by Igor Pro Nika package
}
\examples{
\dontrun{read_nika_sec("nika_sec_HLp_0.dat")}
}
\author{
K. Juraic
}

