% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{read_pilatus_expTime}
\alias{read_pilatus_expTime}
\title{read_pilatu_expTime}
\usage{
read_pilatus_expTime(tifName)
}
\arguments{
\item{tifName}{tiff file name}
}
\value{
exposure time in seconds
}
\description{
read exposure time directly from Pilatus tif image
}
\examples{
\dontrun{read_pilat_expTimeu("saxs_data.tif")}

}
\author{
K. Juraic
}

