% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import_data.R
\name{readFit2dChi}
\alias{readFit2dChi}
\title{readFit2dChi}
\usage{
readFit2dChi(chiName = file.choose())
}
\arguments{
\item{chiName}{chi file name}
}
\value{
datOut data.frame(Q[nm], Intensity[a.u])
}
\description{
read chi file (fit2D integrated SAXS data)
}
\examples{
\dontrun{readFit2dChi("saxs_data.chi")}
}
\author{
K. Juraic
}

