% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{readLogBook}
\alias{readLogBook}
\title{readLogBook( logFile)}
\usage{
readLogBook(logFile = file.choose())
}
\arguments{
\item{logFile}{logbook file name,}
}
\value{
out data.frame(name, angle, time, comment)
}
\description{
read AustroSAXS logbook file. Call without parameter will open file.chose dialog
}
\examples{
 \dontrun{readLogBook("saxs_data.log")}

}
\author{
K. Juraic
}
