% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import_data.R
\name{readLstFit2dChi}
\alias{readLstFit2dChi}
\title{read list of fit2D chi file}
\usage{
readLstFit2dChi(lstFname)
}
\arguments{
\item{lstFname}{list of file names}
}
\value{
data.frame(q,intens)
}
\description{
read fit2D integrated SAXS data
}
\examples{
\dontrun{chi_lst <- c("1.chi","2.chi")
readLstFit2dChi(chi_lst)
}
}
\author{
K. Juraic
}

