% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{readLstNika1d}
\alias{readLstNika1d}
\title{readLstNika1d}
\usage{
readLstNika1d(lstFname)
}
\arguments{
\item{lstFname}{list of file names}
}
\value{
data.frame(q,intens)
}
\description{
read list of NIKA IgorPro integrated SAXS data
}
\examples{
\dontrun{saxs_chi_lst <- c("1.chi","2.chi")
readLstNika1d(saxs_chi_lst)
}
}
\author{
K. Juraic
}
