% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{readNika1d}
\alias{readNika1d}
\title{readNika1d}
\usage{
readNika1d(fname)
}
\arguments{
\item{fname}{data file name}
}
\value{
data.frame(q, intens, err, unknown)
}
\description{
read integrated SAXS 1D data reduced by NIKA IgorPro package
}
\examples{
\dontrun{readNika1d("saxs_data.dat")}

}
\author{
K. Juraic
}

