% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import_data.R
\name{readNika1d_waxs}
\alias{readNika1d_waxs}
\title{readNika1d_waxs}
\usage{
readNika1d_waxs(fname)
}
\arguments{
\item{fname}{data file name}
}
\value{
data.frame(tth, Imean, Ierr, unknown)
}
\description{
read integrated WAXS 1D data reduced by NIKA IgorPro package
}
\examples{
\dontrun{readNika1d_waxs("waxs_data.dat")}
}
\author{
K. Juraic
}

