% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{readNikaCalib}
\alias{readNikaCalib}
\title{Read Nika calibration parameters}
\usage{
readNikaCalib(calibFile)
}
\arguments{
\item{calibFile}{full path for calibration file}
}
\value{
data.frame with calibration data
}
\description{
Readcalibraion parameters calculated by Nika Igor Pro package.
             Parameters shoud be written to tab delimited file. First row
             shoud have parametr names (pxcH	pxcV	saDet	tiltH	tiltV) and
             second row values. If calibration file does not exist program
             asks to insert parameters manualy and then write this
             parameters to file.
}
\examples{
       \dontrun{readNikaCalib(calibFile = "calib.txt")}
}
\author{
Krunoslav Juraic
}

