% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import_data.R
\name{readPilatus1M}
\alias{readPilatus1M}
\title{readPilatus1M}
\usage{
readPilatus1M(fname = file.choose())
}
\arguments{
\item{fname}{tiff file name}
}
\value{
mat image data in 2D matrix
}
\description{
read Pilatus 1M tif file
}
\examples{
\dontrun{readPilatus1M("saxs_data.tif")}
}
\author{
K. Juraic
}

