% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{readPilatus1Mlst}
\alias{readPilatus1Mlst}
\title{readPilatus1Mlst}
\usage{
readPilatus1Mlst(fnames)
}
\arguments{
\item{fnames}{tiff file name list}
}
\value{
list(mat) list of image data in 2D matrix
}
\description{
read list of Pilatus 1M tif file
}
\examples{
\dontrun{readPilatus1Mlist(list)}

}
\author{
K. Juraic
}

