% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\name{sec2file}
\alias{sec2file}
\title{save section data.frame to file}
\usage{
sec2file(sec, fname, colSubset = 1:nrow(sec))
}
\arguments{
\item{sec}{section data frame}

\item{fname}{section file name (secH_430_10_sample_name.dat)}

\item{colSubset}{subset of column to save to file}
}
\description{
save section data.frame to file
}
\examples{
 \dontrun{sec2file(sec, fname)}
}
