% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calibration.R
\name{secCalcAngleQ}
\alias{secCalcAngleQ}
\title{Section Angle and Q scale calculation}
\usage{
secCalcAngleQ(sec, ai, calib)
}
\arguments{
\item{sec}{section data.frame(pxH, pxV, Imean, Isd)}

\item{ai}{angle of incidence in deg.}

\item{calib}{data.frame with calibration parameters (pxcH, pxcV, saDet, tiltH, tiltV)}
}
\value{
data.frame(pxH, pxV, Imean, Isd, aif, tthf, af, Qx, Qy, Qz)
}
\description{
calculate Angular and Q scale for section from pixels array
}
\examples{
\dontrun{sc <- secCalcAngleQ(secDf, ai, calibDf)}
}
\author{
K. Juraic
}

