% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\name{secDfAddAngle}
\alias{secDfAddAngle}
\title{Add angular scale to section}
\usage{
secDfAddAngle(sec, ai, calib)
}
\arguments{
\item{sec}{section data frame (pxH, pxV, Imean, Isd)}

\item{ai}{angle of incidence in deg.}

\item{calib}{calibration parameters (pxcH, pxcV, saDet, tiltH, tiltV)}
}
\value{
data.frame(pxH, pxV, Imean, Isd, aif, ai, tthf)
}
\description{
Add angular scale columns (aif, af, tthf) to section data frame
}
\examples{
 \dontrun{secDfAddAngle(sec, ai, calib)}
}
\author{
K. Juraic
}

