% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/section.R
\name{secDfAddQ}
\alias{secDfAddQ}
\title{Add Q scale to section}
\usage{
secDfAddQ(sec, ai, lambda = 0.154)
}
\arguments{
\item{sec}{section data frame (pxH, pxV, Imean, Isd, aif, ai, tthf)}

\item{ai}{angle of incidence in deg.}

\item{lambda}{wavelength in nm}
}
\value{
data.frame(pxH, pxV, Imean, Isd, aif, ai, tthf, Qx, Qy, Qz, Qabs)
}
\description{
Add Q scale columns (qx, Qy, Qz, Qabs) to section data frame
}
\examples{
\dontrun{secDfAddAngle(sec, ai, lambda)}
}
\author{
K. Juraic
}

