% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/section.R
\name{secLstShow}
\alias{secLstShow}
\title{secLstShow(secLst, titles, ...)}
\usage{
secLstShow(secLst, titles, colSelect = c(1, 3), ...)
}
\arguments{
\item{secLst}{list of matrix sections}

\item{titles}{main title for plots}

\item{colSelect}{columns to plot}

\item{...}{other parameters log='y'}
}
\description{
Show list of matrix sections
}
\examples{
\dontrun{secLstShow(dat, fNames, log='y')}
}
\author{
K. Juraic
}

