% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\name{section_preview}
\alias{section_preview}
\title{preview horizontal and vertical 1D section for 2D GISAXS image}
\usage{
section_preview(gisaxs_mat)
}
\arguments{
\item{gisaxs_mat}{matrix with GISAXS pattern}
}
\value{
section_lst last viewed sections
}
\description{
at specific pixel and withspecific width preview horizontal and
             vertical corssection for 2D GISAXS image
}
\examples{
\dontrun{section_preview(gisaxs.mat)}
}
\author{
K. Juraic
}
