% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\name{sectionH}
\alias{sectionH}
\title{sectionH(mat, pos, width = 10)}
\usage{
sectionH(mat, pos, width = 10)
}
\arguments{
\item{mat}{2D matrix}

\item{pos}{trace position}

\item{width}{trace width}
}
\value{
out.df section data.frame(pxH, pxV, Iraw)
}
\description{
make a horizontal section of 2D matrix at specific position and width
}
\examples{
\dontrun{sectionH(gisaxs.mat)}
}
\author{
K. Juraic
}

