% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/section.R
\name{sectionHlst}
\alias{sectionHlst}
\title{Horizontal section for list of tif images}
\usage{
sectionHlst(tifLst, tifName, pos, width = 10, write2file = FALSE,
  path = "./")
}
\arguments{
\item{tifLst}{list of 2D matrices}

\item{tifName}{names of tif files}

\item{pos}{array of trace position in pixels}

\item{width}{trace width in pixels}

\item{write2file}{store sections to files}

\item{path}{folder where to store sections}
}
\value{
sec list of section list(data.frame(pxH, pxV, Imean, Isd))
}
\description{
make a horizontal section for list of 2D matrix at specific
             positions and width. And save to file
}
\examples{
\dontrun{sectionHlst(tifLst, tifName, pos, width)}
}
\author{
K. Juraic
}

