% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/section.R
\name{sectionV}
\alias{sectionV}
\title{sectionV(mat, pos, width = 10)}
\usage{
sectionV(mat, pos, width = 10)
}
\arguments{
\item{mat}{2D matrix}

\item{pos}{trace position}

\item{width}{trace width}
}
\value{
out.df section data.frame(pxH, pxV, Imean, Isd)
}
\description{
make a vertical section of 2D matrix at specific position and width
}
\examples{
\dontrun{sectionH(gisaxs.mat)}
}
\author{
K. Juraic
}

