% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\name{sectionVlst}
\alias{sectionVlst}
\title{Vertical section for list of tif images}
\usage{
sectionVlst(tifLst, tifName, pos, width = 10, write2file = FALSE,
  path = "./")
}
\arguments{
\item{tifLst}{list of 2D matrices}

\item{tifName}{names of tif files}

\item{pos}{trace position in pixels}

\item{width}{trace width in pixels}

\item{write2file}{store sections to files}

\item{path}{folder where to store sections}
}
\value{
sec list of section list(data.frame(pxH, pxV, Imean, Isd))
}
\description{
make a vertical section for list of 2D matrix at specific
             position and width. And save to file
}
\examples{
\dontrun{sectionVlst(tifLst, tifName, pos, width)}
}
\author{
K. Juraic
}

