% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{tthfCalc}
\alias{tthfCalc}
\title{tthfCalc}
\usage{
tthfCalc(pxH, pxCenterH, samDet, tiltH = 0, pxSizeH = 0.172)
}
\arguments{
\item{pxH}{pixel position (horizontal coordinate)}

\item{pxCenterH}{cordinate of direct beam position}

\item{samDet}{sample to detector distance}

\item{tiltH}{horizontal detector tilt}

\item{pxSizeH}{detector pixel size in horizontal direction}
}
\value{
angle 2*theta exit for pxH
}
\description{
calculate angle 2*theta exit from from known gisaxs
             parameter calibration (central pixel, sample to detector distance)
}
\examples{
tthfCalc(100, 50, 1700, -5.21)
}
\author{
K. Juraic
}

