% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/export_data.R
\name{writeIsGisaxsDat}
\alias{writeIsGisaxsDat}
\title{writeIsGisaxsDat}
\usage{
writeIsGisaxsDat(fName, af, tthf, intens, sigmaI, fitTF,
  weight = rep(1:length(fName)), scaleFac = 1, shiftFac = 0,
  comments = "")
}
\arguments{
\item{fName}{IsGISAXS dat file name}

\item{af}{exit angle alpha.f}

\item{tthf}{exit angle 2 theta}

\item{intens}{intensity}

\item{sigmaI}{intensity error}

\item{fitTF}{iclud this point i fit (T OR f)}

\item{weight}{weigth fot evry point of cross section}

\item{scaleFac}{The scale factor of the cross section with hard bounds below.}

\item{shiftFac}{The shift factor of the cross section with hard bound below.}

\item{comments}{experiment or data comment}
}
\value{
fName
}
\description{
write 1D experimental data in IsGISAXS dat format for fiting
             Experimental data should be given in format
             data.frame(T, sin(2tth), sin(alphaF), Intensity, sigmaI)
}
\examples{
\dontrun{writeIsGisaxsDat(fName = "exp_data.dat", af, tthf, intens, sigmaI, fitTF,
weight = 1, scaleFac = 1, shiftFac = 0, comments = "")}
}
\author{
K. Juraic
}

