% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{writeLogBook}
\alias{writeLogBook}
\title{writeLogBook( logFile)}
\usage{
writeLogBook(logDf, logFile = file.choose())
}
\arguments{
\item{logDf}{logbook data.frame}

\item{logFile}{logbook file name,}
}
\value{
logbook file name
}
\description{
write AustroSAXS logbook file.
}
\examples{
 \dontrun{writeLogBook(logDf)}
}
\author{
K. Juraic
}

