% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.R
\name{writeP00}
\alias{writeP00}
\title{writeP00}
\usage{
writeP00(fname, qExp, iExp, distSD, pxC, pxSize, pxN, lambda)
}
\arguments{
\item{fname}{file name}

\item{qExp}{Q vector}

\item{iExp}{intensity values}

\item{distSD}{sample to detector distance}

\item{pxC}{direct beam position in px}

\item{pxSize}{detector pixel size}

\item{pxN}{number of pixels}

\item{lambda}{wavelength}
}
\value{
out data.frame(px,intens)
}
\description{
export 1D SAXs data to p00 format
}
\examples{
write\dontrun{P00(fname, qExp, iExp, distSD, pxC, pxSize, pxN, lambda)}

}
\author{
K. Juraic
}

