% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisaxs_model.R
\name{Yoneda}
\alias{Yoneda}
\title{Yoneda(tthf, af, ai, delta, beta, pI0, pF, pE, lambda = .154)}
\usage{
Yoneda(tthf, af, ai, delta, beta, surfRms = 0, pI0 = 0, pF = 1, pE = 1,
  lambda = 0.154)
}
\arguments{
\item{tthf}{scattering angle out of specular plane}

\item{af}{scattering angle in specular plane}

\item{ai}{angle of incidence}

\item{delta}{index of refraction (Im part)}

\item{beta}{index of refractong (Re part)}

\item{surfRms}{surface roughness in [nm]}

\item{pI0}{shift factor}

\item{pF}{scaling factor}

\item{pE}{exponent}

\item{lambda}{X-ray wavelength in [nm]}
}
\value{
intensity value
}
\description{
calculate intensity arround Yoneda maksimum
}
\examples{
Yoneda(tthf = 0, af = 1.5, ai = .22,
       delta = 7e-5, beta = 5e-7,
       surfRms = 1,
       pI0 = 0, pF = 1, pE = 3)
}
\author{
K. Juraic
}
